Move busybox helpers to their own file
These are referenced in at least two different places, so put them in
busybox.go for slightly easier discovery.
